fix(lint): honor resident config cache opt-out
