+++ packages/lint/test/config/windows_junction_treats_paths_as_data_test.go
@@ -1,8 +1,6 @@
 package linthost
 
 import (
-  "crypto/sha256"
-  "encoding/hex"
   "os"
   "path/filepath"
   "runtime"
@@ -40,38 +38,5 @@ func TestWindowsJunctionTreatsPathsAsData(t *testing.T) {
   expanded := strings.ReplaceAll(link, "%TTSC_WINDOWS_JUNCTION_PROBE%", "expanded")
   if _, err := os.Lstat(expanded); !os.IsNotExist(err) {
     t.Fatalf("cmd.exe expanded a percent sequence into %s: %v", expanded, err)
-  }
-}
-
-func TestWindowsJunctionDependencyEntryMatchesNodeFingerprint(t *testing.T) {
-  if runtime.GOOS != "windows" {
-    t.Skip("Windows junction boundary")
-  }
-  root := t.TempDir()
-  target := filepath.Join(root, "target")
-  link := filepath.Join(root, "link")
-  if err := os.MkdirAll(target, 0o755); err != nil {
-    t.Fatal(err)
-  }
-  if err := windowsjunction.Create(link, target); err != nil {
-    t.Fatal(err)
-  }
-  linkTarget, err := os.Readlink(link)
-  if err != nil {
-    t.Fatalf("read junction target: %v", err)
-  }
-  h := sha256.New()
-  h.Write([]byte("symlink\x00"))
-  h.Write([]byte(linkTarget))
-  want := hex.EncodeToString(h.Sum(nil))
-  got, err := configDependencyDigest(configDependencyFingerprint{
-    Path: link,
-    Kind: configDependencyEntry,
-  })
-  if err != nil {
-    t.Fatalf("digest junction entry: %v", err)
-  }
-  if got != want {
-    t.Fatalf("junction entry digest = %q, want Node symlink digest %q", got, want)
   }
 }
