+++ website/src/content/docs/development/concepts/protocol.mdx
@@ -261,7 +261,7 @@ A sidecar that does not implement `lsp-serve` rejects the subcommand, and the ho
 
 `ttscserver` is not the only client of this stream. `@ttsc/graph` opens one daemon per plugin binary for the life of a resident graph session and asks `project-inputs` and `graph-nodes` over it. It re-derives the artifact set whenever a document or lint configuration behind that set moves, and a process, a plugin load, and a configuration evaluation per edit is what the daemon exists to avoid. Neither verb carries a `uri`, because both describe the project rather than a document, as `lsp-hints` does. That client sends `{ "invalidate": true }` on the first verb of each republish: which sources exist is what activates a rule's claims, and the developer has been editing code as well as documents.
 
-`@ttsc/lint` keeps its resolved rule configuration in the daemon independently of the Program. Reuse is authorized only while the executable config loader's complete dependency fingerprints remain current, including missing resolution candidates, directory topology, physical path identity, and package implementation files. Package implementation files remain cache inputs rather than project-input watch outputs, so this validation does not widen editor watches into `node_modules`.
+`@ttsc/lint` keeps its resolved rule configuration in the daemon independently of the Program. Reuse is authorized only while the executable config loader's complete dependency fingerprints remain current, including missing resolution candidates, directory topology, physical path identity, and package implementation files. Package implementation files remain cache inputs rather than project-input watch outputs, so this validation does not widen editor watches into `node_modules`. `TTSC_LINT_DISABLE_CONFIG_CACHE=1` disables both executable-config evaluation caching and this resident memo.
 
 Both verbs join the daemon on the weaker condition above, and the fallback matters more for them than for a diagnostic. An empty artifact set is the correct answer for most projects, so a daemon that quietly answered nothing would be indistinguishable from a project that publishes nothing. A nonzero reply therefore means "ask the direct command", never "there are none".
